+++ src/types/nitro.ts
@@ -7,7 +7,6 @@ import type { Consola, LogLevel } from 'consola'
 import { WatchOptions } from 'chokidar'
 import type { RollupCommonJSOptions } from '@rollup/plugin-commonjs'
 import type { NodeExternalsOptions } from '../rollup/plugins/externals'
-import type { StorageMounts } from '../rollup/plugins/storage'
 import type { RollupConfig } from '../rollup/config'
 import type { Options as EsbuildOptions } from '../rollup/plugins/esbuild'
 import { NitroErrorHandler, NitroDevEventHandler, NitroEventHandler } from './handler'
@@ -28,6 +27,13 @@ export interface NitroHooks {
   'compiled': (nitro: Nitro) => HookResult
   'dev:reload': () => HookResult
   'close': () => HookResult
+}
+
+export interface StorageMounts {
+  [path: string]: {
+    driver: 'fs' | 'http' | 'memory' | 'redis' | 'cloudflare-kv',
+    [option: string]: any
+  }
 }
 
 type DeepPartial<T> = T extends Record<string, any> ? { [P in keyof T]?: DeepPartial<T[P]> | T[P] } : T
@@ -94,6 +100,8 @@ export interface NitroOptions {
 
   // Features
   storage: StorageMounts
+  devStorage: StorageMounts
+  bundledStorage: string[]
   timing: boolean
   renderer: string
   serveStatic: boolean
@@ -134,7 +142,7 @@ export interface NitroOptions {
   node: boolean
   moduleSideEffects: string[]
   esbuild?: {
-    options?: EsbuildOptions
+    options?: Partial<EsbuildOptions>
   }
   noExternals: boolean,
   externals: NodeExternalsOptions